requirements.txt
# TinyTroupe Service Requirements
# Flask and extensions
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23

# CLI
click==8.1.7
tabulate==0.9.0
requests==2.31.0

# TinyTroupe dependencies
# Note: TinyTroupe should be installed separately from GitHub
# pip install git+https://github.com/microsoft/TinyTroupe.git

# OpenAI for TinyTroupe
openai==1.3.0

# Testing
pytest==7.4.3
